refactor(editor): table overlay controller 분리
