+++ front/e2e/editor-studio-state.spec.ts
@@ -339,6 +339,22 @@ test.describe("editor studio state", () => {
     const blockEditorEngineLayersSource = existsSync(blockEditorEngineLayersPath)
       ? readFileSync(blockEditorEngineLayersPath, "utf8")
       : ""
+    const blockEditorTableOverlayLayerPath = path.resolve(
+      __dirname,
+      "../src/components/editor/BlockEditorEngine.tableOverlayLayer.tsx"
+    )
+    expect(existsSync(blockEditorTableOverlayLayerPath)).toBe(true)
+    const blockEditorTableOverlayLayerSource = existsSync(blockEditorTableOverlayLayerPath)
+      ? readFileSync(blockEditorTableOverlayLayerPath, "utf8")
+      : ""
+    const blockEditorTableOverlayControllerPath = path.resolve(
+      __dirname,
+      "../src/components/editor/useBlockEditorTableOverlayController.ts"
+    )
+    expect(existsSync(blockEditorTableOverlayControllerPath)).toBe(true)
+    const blockEditorTableOverlayControllerSource = existsSync(blockEditorTableOverlayControllerPath)
+      ? readFileSync(blockEditorTableOverlayControllerPath, "utf8")
+      : ""
     const blockSelectionModelSource = readFileSync(
       path.resolve(__dirname, "../src/components/editor/blockSelectionModel.ts"),
       "utf8"
@@ -862,9 +878,13 @@ test.describe("editor studio state", () => {
     expect(blockEditorShellSource).toContain('import BlockEditorEngine from "./BlockEditorEngine"')
     expect(blockEditorShellSource).toContain('import type { BlockEditorEngineProps } from "./blockEditorEngineTypes"')
     expect(blockEditorShellSource).toContain("const BlockEditorShell = (props: BlockEditorEngineProps) => <BlockEditorEngine {...props} />")
-    expect(blockEditorEngineSource.split("\n").length).toBeLessThanOrEqual(7000)
+    expect(blockEditorEngineSource.split("\n").length).toBeLessThanOrEqual(5500)
     expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.styles"')
     expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.layers"')
+    expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.tableOverlayLayer"')
+    expect(blockEditorEngineSource).toContain('} from "./useBlockEditorTableOverlayController"')
+    expect(blockEditorTableOverlayLayerSource).toContain("export const BlockEditorTableOverlayLayer =")
+    expect(blockEditorTableOverlayControllerSource).toContain("export const useBlockEditorTableOverlayController =")
     expect(blockEditorEngineLayersSource).toContain("export const BlockEditorToolbarLayer =")
     expect(blockEditorEngineLayersSource).toContain("export const BlockEditorQuickInsertLayer =")
     expect(blockEditorEngineLayersSource).toContain("export const BlockEditorSlashMenuLayer =")
@@ -873,6 +893,7 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineLayersSource).toContain('data-testid="slash-menu"')
     expect(blockEditorEngineLayersSource).toContain('data-testid="editor-text-bubble-toolbar"')
     expect(blockEditorEngineLayersSource).toContain('"block-drag-handle"')
+    expect(blockEditorEngineLayersSource).not.toContain("export const BlockEditorTableOverlayLayer =")
     expect(blockEditorEngineSource).not.toContain("<ToolbarActions>")
     expect(blockEditorEngineSource).not.toContain("<QuickInsertBar")
     expect(blockEditorEngineSource).not.toContain("<SlashMenu")
@@ -993,7 +1014,7 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toMatch(/const\s+isTabBlockSelectionEligible\s*=/)
     expect(blockEditorEngineSource).toContain('from "./blockHandleLayoutModel"')
     expect(blockEditorEngineSource).toContain('from "./blockDragModel"')
-    expect(blockEditorEngineSource).toContain('from "./tableAxisDragModel"')
+    expect(blockEditorTableOverlayControllerSource).toContain('from "./tableAxisDragModel"')
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveBlockHandleAnchorTop\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveThinBlockHandleAnchorTop\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveBlockHandleRailLayout\s*=/)
@@ -1266,6 +1287,14 @@ test.describe("editor studio state", () => {
       path.resolve(__dirname, "../src/components/editor/BlockEditorEngine.layers.tsx"),
       "utf8"
     )
+    const blockEditorTableOverlayLayerSource = readFileSync(
+      path.resolve(__dirname, "../src/components/editor/BlockEditorEngine.tableOverlayLayer.tsx"),
+      "utf8"
+    )
+    const blockEditorTableOverlayControllerSource = readFileSync(
+      path.resolve(__dirname, "../src/components/editor/useBlockEditorTableOverlayController.ts"),
+      "utf8"
+    )
     const tableAffordanceModelSource = readFileSync(
       path.resolve(__dirname, "../src/components/editor/tableAffordanceModel.ts"),
       "utf8"
@@ -1331,9 +1360,9 @@ test.describe("editor studio state", () => {
       "utf8"
     )
 
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-column-drag-guide"')
-    expect(blockEditorEngineLayersSource).toContain('data-testid={`table-column-resize-boundary-${index}`}')
-    expect(blockEditorEngineSource).toContain('} from "./tableResizeInteractionModel"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-column-drag-guide"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid={`table-column-resize-boundary-${index}`}')
+    expect(blockEditorTableOverlayControllerSource).toContain('} from "./tableResizeInteractionModel"')
     expect(tableResizeInteractionModelSource).toContain("export type TableRowResizeState = {")
     expect(tableResizeInteractionModelSource).toContain("export type TableColumnRailResizeState = {")
     expect(tableResizeInteractionModelSource).toContain("export type TableColumnDragGuideState = {")
@@ -1369,21 +1398,21 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toContain("const isRowResizeHandleTarget = useCallback(")
     expect(blockEditorEngineSource).not.toContain("const syncTableColumnDragGuideForColumn = useCallback(")
     expect(blockEditorEngineSource).not.toContain("const startTableColumnResizeFromDomHandle = useCallback(")
-    expect(blockEditorEngineSource).toContain('} from "./tableFloatingUiModel"')
-    expect(blockEditorEngineSource).toContain("const getActiveTableRectFromDom = useCallback(")
-    expect(blockEditorEngineLayersSource).toContain('import { createPortal } from "react-dom"')
-    expect(blockEditorEngineLayersSource).toContain("const tableOverlay = (")
-    expect(blockEditorEngineLayersSource).toContain("createPortal(tableOverlay, document.body)")
-    expect(blockEditorEngineSource).toContain("const TABLE_EDGE_HANDLE_INSET_PX = 6")
+    expect(blockEditorTableOverlayControllerSource).toContain('} from "./tableFloatingUiModel"')
+    expect(blockEditorTableOverlayControllerSource).toContain("const getActiveTableRectFromDom = useCallback(")
+    expect(blockEditorTableOverlayLayerSource).toContain('import { createPortal } from "react-dom"')
+    expect(blockEditorTableOverlayLayerSource).toContain("const tableOverlay = (")
+    expect(blockEditorTableOverlayLayerSource).toContain("createPortal(tableOverlay, document.body)")
+    expect(blockEditorTableOverlayControllerSource).toContain("const TABLE_EDGE_HANDLE_INSET_PX = 6")
     expect(tableAffordanceModelSource).toContain("export const TABLE_EDGE_ADD_BUTTON_SIZE_PX = 24")
     expect(tableAffordanceModelSource).toContain("export const TABLE_CELL_MENU_BUTTON_SIZE_PX = 22")
     expect(tableAffordanceModelSource).toContain("export const TABLE_ADD_BAR_VIEWPORT_PADDING_PX = 8")
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-corner-grow-handle"')
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-corner-preview-outline"')
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-structure-menu-button"')
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-cell-menu-button"')
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-overflow-mode-normal"')
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-overflow-mode-wide"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-corner-grow-handle"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-corner-preview-outline"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-structure-menu-button"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-cell-menu-button"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-overflow-mode-normal"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-overflow-mode-wide"')
     expect(tableWidthModelSource).toContain('export const TABLE_OVERFLOW_MODE_WIDE = "wide"')
     expect(tableWidthModelSource).toContain("export const getTableOverflowMode = (")
     expect(tableWidthModelSource).toContain("export const shouldPromoteWideTableOverflowMode = (")
@@ -1441,74 +1470,76 @@ test.describe("editor studio state", () => {
     expect(tableRenderedDomModelSource).toContain("export const resolveTableScopedSelectedCell = (")
     expect(tableRenderedDomModelSource).toContain("export const resolveActiveRenderedTableForFloatingUi = (")
     expect(tableRenderedDomModelSource).toContain("export const readRenderedColumnWidths = (")
-    expect(blockEditorEngineSource).toContain('} from "./tableRenderedDomModel"')
+    expect(blockEditorTableOverlayControllerSource).toContain('} from "./tableRenderedDomModel"')
     expect(blockEditorEngineSource).not.toContain("const findActiveRenderedTable = (")
     expect(blockEditorEngineSource).not.toContain("const resolveTableScopedSelectedCell = (")
     expect(blockEditorEngineSource).not.toContain("const resolveActiveRenderedTableForFloatingUi = (")
     expect(blockEditorEngineSource).not.toContain("const readRenderedColumnWidths = (")
-    expect(blockEditorEngineLayersSource).toContain('data-testid="table-row-drag-shadow"')
-    expect(blockEditorEngineLayersSource).toContain('"table-row-reorder-indicator"')
-    expect(blockEditorEngineLayersSource).toContain('"table-column-reorder-indicator"')
+    expect(blockEditorTableOverlayLayerSource).toContain('data-testid="table-row-drag-shadow"')
+    expect(blockEditorTableOverlayLayerSource).toContain('"table-row-reorder-indicator"')
+    expect(blockEditorTableOverlayLayerSource).toContain('"table-column-reorder-indicator"')
     expect(blockEditorEngineStylesSource).toContain('const TableCellMenuButton = styled(TableHandleButton)`')
-    expect(blockEditorEngineSource).toContain("const updateActiveTableOverflowMode = useCallback(")
-    expect(blockEditorEngineSource).toContain("const reorderTableAxisAtPosition = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const updateActiveTableOverflowMode = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const reorderTableAxisAtPosition = useCallback(")
     expect(tableCornerGrowModelSource).toContain("export type TableCornerGrowState = {")
     expect(tableCornerGrowModelSource).toContain("export type TableCornerPreviewState = {")
     expect(tableCornerGrowModelSource).toContain("export const resolveTableCornerPreviewState = (")
     expect(tableCornerGrowModelSource).toContain("export const resolveTableCornerGrowStepMetrics = (")
     expect(tableCornerGrowModelSource).toContain("export const resolveTableCornerGrowStepMetricsFromDataset = (")
-    expect(blockEditorEngineSource).toContain('} from "./tableCornerGrowModel"')
+    expect(blockEditorTableOverlayControllerSource).toContain('} from "./tableCornerGrowModel"')
     expect(blockEditorEngineSource).not.toContain("type TableCornerGrowState = {")
     expect(blockEditorEngineSource).not.toContain("type TableCornerPreviewState = {")
     expect(blockEditorEngineSource).not.toContain("const resolveTableCornerPreviewState = useCallback(")
     expect(blockEditorEngineSource).not.toContain("const getTableCornerGrowStepMetrics = useCallback(")
     expect(blockEditorEngineSource).not.toContain("const getTableCornerGrowStepMetricsFromHandle = useCallback(")
-    expect(blockEditorEngineSource).toContain("const applyTableCornerGrowSteps = useCallback(")
-    expect(blockEditorEngineSource).toContain("const shrinkTableAxisAtEnd = useCallback(")
-    expect(blockEditorEngineSource).toContain("const beginTableAxisDragFromPending = useCallback(")
-    expect(blockEditorEngineSource).toContain("const startPendingTableAxisDrag = useCallback(")
-    expect(blockEditorEngineSource).toContain("const selectTableAxisAtIndex = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const applyTableCornerGrowSteps = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const shrinkTableAxisAtEnd = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const beginTableAxisDragFromPending = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const startPendingTableAxisDrag = useCallback(")
+    expect(blockEditorTableOverlayControllerSource).toContain("const selectTableAxisAtIndex = useCallback(")
     expect(tableStructureModelSource).toContain('overflowMode: getTableOverflowMode(tableNode)')
     expect(tableWidthModelSource).toContain("const maxActiveWidth = Math.max(TABLE_MIN_COLUMN_WIDTH_PX, safeBudget - otherColumnsWidth)")
-    expect(blockEditorEngineSource).toContain("const tableCornerGrowSuppressClickRef = useRef(false)")
+    expect(blockEditorTableOverlayControllerSource).toContain("const tableCornerGrowSuppressClickRef = useRef(false)")
     expect(blockEditorEngineStylesSource).toContain('"grip" | "grow"')
-    expect(blockEditorEngineSource).toContain("isCellMenuOpen,")
+    expect(blockEditorTableOverlayControllerSource).toContain("isCellMenuOpen,")
     expect(markdownRendererSource).toContain("const explicitTableWidth = useMemo(")
     expect(markdownRendererSource).toContain("minWidth: `${explicitTableWidth}px`")
     expect(markdownRendererRootSource).toContain("width: auto;")
     expect(markdownRendererRootSource).toContain("max-width: none;")
-    expect(blockEditorEngineSource).toContain("isRowMenuOpen,")
-    expect(blockEditorEngineSource).toContain("isColumnMenuOpen,")
-    expect(blockEditorEngineLayersSource).toContain("activeTableStructureState.hasHeaderRow")
-    expect(blockEditorEngineLayersSource).toContain("activeTableStructureState.hasHeaderColumn")
-    expect(blockEditorEngineSource).toContain("const tableCornerGrowStepMetrics = resolveTableCornerGrowStepMetrics(tableAffordanceGeometry)")
-    expect(blockEditorEngineLayersSource).toContain("data-column-step={tableCornerGrowStepMetrics.columnStepPx}")
-    expect(blockEditorEngineLayersSource).toContain("data-row-step={tableCornerGrowStepMetrics.rowStepPx}")
-    expect(blockEditorEngineLayersSource).toContain('aria-label="표 구조 메뉴"')
-    expect(blockEditorEngineLayersSource).toContain("페이지 너비에 맞춤")
-    expect(blockEditorEngineLayersSource).toContain("넓은 표")
-    expect(blockEditorEngineLayersSource).toContain("제목 행")
-    expect(blockEditorEngineLayersSource).toContain("제목 열")
+    expect(blockEditorTableOverlayControllerSource).toContain("isRowMenuOpen,")
+    expect(blockEditorTableOverlayControllerSource).toContain("isColumnMenuOpen,")
+    expect(blockEditorTableOverlayLayerSource).toContain("activeTableStructureState.hasHeaderRow")
+    expect(blockEditorTableOverlayLayerSource).toContain("activeTableStructureState.hasHeaderColumn")
+    expect(blockEditorTableOverlayControllerSource).toContain("const tableCornerGrowStepMetrics = resolveTableCornerGrowStepMetrics(tableAffordanceGeometry)")
+    expect(blockEditorTableOverlayLayerSource).toContain("data-column-step={tableCornerGrowStepMetrics.columnStepPx}")
+    expect(blockEditorTableOverlayLayerSource).toContain("data-row-step={tableCornerGrowStepMetrics.rowStepPx}")
+    expect(blockEditorTableOverlayLayerSource).toContain('aria-label="표 구조 메뉴"')
+    expect(blockEditorTableOverlayLayerSource).toContain("페이지 너비에 맞춤")
+    expect(blockEditorTableOverlayLayerSource).toContain("넓은 표")
+    expect(blockEditorTableOverlayLayerSource).toContain("제목 행")
+    expect(blockEditorTableOverlayLayerSource).toContain("제목 열")
     expect(tableAffordanceModelSource).toContain("export type TableAffordanceGeometry = {")
     expect(tableAffordanceModelSource).toContain("export type TableAffordanceVisibility = {")
     expect(tableAffordanceModelSource).toContain("export const INITIAL_TABLE_AFFORDANCE_GEOMETRY: TableAffordanceGeometry = {")
     expect(tableAffordanceModelSource).toContain("export const INITIAL_TABLE_AFFORDANCE_VISIBILITY: TableAffordanceVisibility = {")
     expect(tableAffordanceModelSource).toContain("export const resolveDesktopTableRailLayout = (")
-    expect(blockEditorEngineSource).toContain('} from "./tableAffordanceModel"')
+    expect(blockEditorTableOverlayControllerSource).toContain('} from "./tableAffordanceModel"')
     expect(blockEditorEngineSource).not.toContain("type TableAffordanceGeometry = {")
     expect(blockEditorEngineSource).not.toContain("type TableAffordanceVisibility = {")
     expect(blockEditorEngineSource).not.toContain("const INITIAL_TABLE_AFFORDANCE_GEOMETRY: TableAffordanceGeometry = {")
     expect(blockEditorEngineSource).not.toContain("const INITIAL_TABLE_AFFORDANCE_VISIBILITY: TableAffordanceVisibility = {")
-    expect(blockEditorEngineSource).toContain("const [tableAffordanceGeometry, setTableAffordanceGeometry] = useState<TableAffordanceGeometry>(")
-    expect(blockEditorEngineSource).toContain(
+    expect(blockEditorTableOverlayControllerSource).toContain("const [tableAffordanceGeometry, setTableAffordanceGeometry] = useState<TableAffordanceGeometry>(")
+    expect(blockEditorTableOverlayControllerSource).toContain(
       "const [tableAffordanceVisibility, setTableAffordanceVisibility] = useState<TableAffordanceVisibility>("
     )
-    expect(blockEditorEngineSource).toContain("const tableAffordanceVisibilityRef = useRef(tableAffordanceVisibility)")
+    expect(blockEditorTableOverlayControllerSource).toContain("const tableAffordanceVisibilityRef = useRef(tableAffordanceVisibility)")
     expect(blockEditorEngineSource).not.toContain("type TableQuickRailState =")
-    expect(blockEditorEngineSource).toContain("isTableStructureMenuOpen,")
-    expect(blockEditorEngineSource).toContain("shouldShowColumnAddBar,")
-    expect(blockEditorEngineSource).toContain("shouldShowRowAddBar,")
-    expect(blockEditorEngineSource).toContain("findActiveRenderedTable(viewportRef.current, tableAffordanceGeometryRef.current)")
+    expect(blockEditorTableOverlayControllerSource).toContain("isTableStructureMenuOpen,")
+    expect(blockEditorTableOverlayControllerSource).toContain("shouldShowColumnAddBar,")
+    expect(blockEditorTableOverlayControllerSource).toContain("shouldShowRowAddBar,")
+    expect(blockEditorTableOverlayControllerSource).toContain(
+      "findActiveRenderedTable(viewportRef.current, tableAffordanceGeometryRef.current)"
+    )
     expect(blockEditorEngineStylesSource).toContain("display: none !important;")
     expect(editorStudioSource).toContain('import { useEditorStudioPersistence } from "./useEditorStudioPersistence"')
     expect(editorStudioPersistenceSource).toContain("const currentPostContent = postContentLiveRef.current")
